@@ -1,12 +1,14 @@
 {
   "name": "@next2d/framework",
-  "description": "It is a framework dedicated to Next2D that enables scene management by URL (SPA), which has been difficult with conventional Canvas/WebGL applications, and simplifies readability and shareability by fixing the development pattern (MVVM).",
-  "version": "1.6.2",
+  "description": "Next2D Framework is designed according to the principles of clean architecture, domain-driven development, test-driven development, and MVVM, with an emphasis on flexibility, scalability, and maintainability, and a design methodology that keeps each layer loosely coupled.",
+  "version": "2.0.0",
   "homepage": "https://next2d.app",
   "bugs": "https://github.com/Next2D/Framework/issues/new",
-  "author": "Toshiyuki Ienaga <ienaga@tvon.jp> (https://github.com/ienaga/)",
+  "author": "Toshiyuki Ienaga <ienaga@next2d.app> (https://github.com/ienaga/)",
   "license": "MIT",
   "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "type": "module",
   "keywords": [
     "Next2D",
     "Next2D Framework"
@@ -14,10 +16,8 @@
   "scripts": {
     "lint": "eslint src/**/*.ts",
     "publish": "node ./scripts/version.js && tsc",
-    "test": "jest",
-    "jsdoc": "tsc && jsdoc -c jsdoc.conf.js -r dist DOCS.md"
+    "test": "npx vitest"
   },
-  "types": "dist",
   "files": [
     "dist"
   ],
@@ -25,26 +25,14 @@
     "type": "git",
     "url": "git+https://github.com/Next2D/Framework.git"
   },
-  "dependencies": {
-    "@next2d/player": "*"
-  },
   "devDependencies": {
-    "@babel/core": "^7.22.5",
-    "@babel/plugin-transform-modules-commonjs": "^7.22.5",
-    "@babel/preset-env": "^7.22.5",
-    "@next2d/jsdoc-template": "^1.0.7",
-    "@types/jest": "^29.5.2",
-    "@typescript-eslint/eslint-plugin": "^5.60.0",
-    "@typescript-eslint/parser": "^5.60.0",
-    "eslint": "^8.43.0",
-    "eslint-webpack-plugin": "^4.0.1",
-    "jest": "^29.5.0",
-    "jsdoc": "^4.0.2",
-    "ts-jest": "^29.1.0",
-    "ts-loader": "^9.4.3",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.1.3",
-    "webpack": "^5.88.0",
-    "webpack-cli": "^5.1.4"
+    "@next2d/player": "*",
+    "@typescript-eslint/eslint-plugin": "^6.9.1",
+    "@typescript-eslint/parser": "^6.9.1",
+    "eslint": "^8.53.0",
+    "jsdom": "^22.1.0",
+    "typescript": "^5.2.2",
+    "vite": "^4.5.0",
+    "vitest": "^0.34.6"
   }
 }
