@@ -1,85 +1,55 @@
-import { RequestType } from "../../infrastructure/constant/RequestType";
-import { RequestParser } from "../../domain/parser/RequestParser";
+import { execute as requestParser } from "../../domain/parser/RequestParser";
 import { loaderInfoMap } from "../variable/LoaderInfoMap";
 import { response } from "../variable/Response";
 import type { LoaderInfo } from "@next2d/display";
 import type { ParentImpl } from "@next2d/interface";
-
-interface Object {
-    type: string;
-    name: string;
-    path: string;
-    cache?: boolean;
-    class?: string;
-    access?: string;
-    method?: string;
-    callback?: string | string[];
-}
+import { RequestImpl } from "src/interface/RequestImpl";
 
 /**
- * @class
- * @memberof application.service
+ * @param  {string} name
+ * @return {void}
+ * @method
+ * @public
  */
-export class RemoveResponse
+export const execute = (name: string): void =>
 {
-    private readonly _$requestParser: RequestParser;
+    const requests: RequestImpl[] = requestParser(name);
+    for (let idx: number = 0; idx < requests.length; ++idx) {
 
-    /**
-     * @constructor
-     * @public
-     */
-    constructor ()
-    {
+        const object: RequestImpl = requests[idx];
+
+        if (object.type !== "content") {
+            continue;
+        }
+
+        if (object.cache
+            || !object.name
+            || !response.has(object.name)
+        ) {
+            continue;
+        }
+
         /**
-         * @type {RequestParser}
-         * @private
+         * キャッシュしないパッケージはインメモリから削除
+         * Remove non-cached packages from in-memory
          */
-        this._$requestParser = new RequestParser();
-    }
-
-    /**
-     * @param  {string} name
-     * @return {void}
-     * @method
-     * @public
-     */
-    execute (name: string): void
-    {
-        const requests: Object[] = this._$requestParser.execute(name);
-        for (let idx: number = 0; idx < requests.length; ++idx) {
-
-            const object: Object = requests[idx];
-
-            if (object.type !== RequestType.CONTENT) {
-                continue;
-            }
-
-            if (object.cache || !response.has(object.name)) {
-                continue;
-            }
-
-            /**
-             * キャッシュしないパッケージはインメモリから削除
-             * Remove non-cached packages from in-memory
-             */
-            const content: ParentImpl<any> = response.get(object.name);
-            const contentLoaderInfo: LoaderInfo | null = content._$loaderInfo;
-            if (contentLoaderInfo && contentLoaderInfo._$data) {
-                const symbols: Map<string, any> = contentLoaderInfo._$data.symbols;
-                if (symbols.size) {
-                    for (const name of symbols.keys()) {
-                        loaderInfoMap.delete(name);
-                    }
+        const content: ParentImpl<any> = response.get(object.name);
+        const contentLoaderInfo: LoaderInfo | null = content._$loaderInfo;
+        if (contentLoaderInfo && contentLoaderInfo._$data) {
+            const symbols: Map<string, any> = contentLoaderInfo._$data.symbols;
+            if (symbols.size) {
+                for (const name of symbols.keys()) {
+                    loaderInfoMap.delete(name);
                 }
             }
         }
-
-        /**
-         * レスポンスデータをマップに格納
-         * Stores response data in a map
-         */
-        if (response.size) {
-            response.clear();
-        }
     }
-}
+
+    /**
+     * レスポンスデータを初期化
+     * Initialize response data
+     */
+    if (response.size) {
+        response.clear();
+    }
+};
