@@ -1,56 +1,40 @@
-import { parser } from "../../application/variable/Parser";
+import type { RequestImpl } from "src/interface/RequestImpl";
 import { packages } from "../../application/variable/Packages";
 
-interface Object {
-    type: string;
-    name: string;
-    path: string;
-    cache?: boolean;
-    class: string;
-    access: string;
-    method: string;
-    callback?: string|string[];
-    body?: object;
-    headers?: HeadersInit;
-}
-
 /**
- * 外部データ取得時のリクエストとレスポンスの管理クラス
- * Request and Response management class for JSON acquisition
+ * @description 指定先の外部データを非同期で取得
+ *              Asynchronous acquisition of external data at specified destination
  *
- * @class
- * @memberof infrastructure.repository
+ * @param  {object} request_object
+ * @return {Promise<any>}
+ * @method
+ * @public
  */
-export class CustomRepository
+export const execute = (request_object: RequestImpl): Promise<any> =>
 {
-    /**
-     * @description 指定先の外部データを非同期で取得
-     *              Asynchronous acquisition of external data at specified destination
-     *
-     * @param  {object} object
-     * @return {Promise<any>}
-     * @method
-     * @public
-     */
-    execute (object: Object): Promise<any>
+    return new Promise((resolve) =>
     {
-        return new Promise((resolve) =>
-        {
-            const className: string = parser.execute(object.class);
-            if (!packages.has(className)) {
-                return resolve(null);
-            }
+        if (!request_object.class
+            || !request_object.access
+            || !request_object.method
+        ) {
+            return resolve(null);
+        }
 
-            const CallbackClass: any = packages.get(className);
-            const promise: Promise<any> = parser.execute(object.access) === "static"
-                ? Promise.resolve(CallbackClass[parser.execute(object.method)]())
-                : Promise.resolve(new CallbackClass()[parser.execute(object.method)]());
+        const name: string = request_object.class;
+        if (!name || !packages.has(name)) {
+            return resolve(null);
+        }
 
-            return promise
-                .then((value: any) =>
-                {
-                    return resolve(value);
-                });
-        });
-    }
-}
+        const CallbackClass: any = packages.get(name);
+        const promise: Promise<any> = request_object.access === "static"
+            ? Promise.resolve(CallbackClass[request_object.method]())
+            : Promise.resolve(new CallbackClass()[request_object.method]());
+
+        return promise
+            .then((value: any) =>
+            {
+                return resolve(value);
+            });
+    });
+};
