@@ -1,59 +1,39 @@
-import { parser } from "../../application/variable/Parser";
-
-interface Object {
-    type: string;
-    name: string;
-    path: string;
-    cache?: boolean;
-    callback?: string | string[];
-    method?: string;
-    body?: object;
-    headers?: HeadersInit;
-}
+import type { RequestImpl } from "src/interface/RequestImpl";
 
 /**
- * JSON取得時のリクエストとレスポンスの管理クラス
- * Request and Response management class for JSON acquisition
+ * @description 指定先のJSONを非同期で取得
+ *              Asynchronously obtain JSON of the specified destination
  *
- * @class
- * @memberof infrastructure.repository
+ * @param  {object} request_object
+ * @return {Promise}
+ * @method
+ * @public
  */
-export class JsonRepository
+export const execute = async (request_object: RequestImpl): Promise<any> =>
 {
-    /**
-     * @description 指定先のJSONを非同期で取得
-     *              Asynchronously obtain JSON of the specified destination
-     *
-     * @param  {object} object
-     * @return {Promise}
-     * @method
-     * @public
-     */
-    execute (object: Object): Promise<any>
-    {
-        const options: RequestInit = {};
-
-        const method: string = options.method = object.method
-            ? parser.execute(object.method).toUpperCase()
-            : "GET";
-
-        const body: any = object.body
-            && method === "POST" || method === "PUT"
-            ? JSON.stringify(object.body)
-            : null;
-
-        if (body) {
-            options.body = body;
-        }
-
-        if (object.headers) {
-            options.headers = object.headers;
-        }
-
-        return fetch(`${parser.execute(object.path)}`, options)
-            .then((response: Response) =>
-            {
-                return response.json();
-            });
+    if (!request_object.path) {
+        throw new Error("`path` must be set for json requests.");
     }
-}
+
+    const options: RequestInit = {};
+
+    const method: string = options.method = request_object.method
+        ? request_object.method.toUpperCase()
+        : "GET";
+
+    const body: any = request_object.body
+        && method === "POST" || method === "PUT"
+        ? JSON.stringify(request_object.body)
+        : null;
+
+    if (body) {
+        options.body = body;
+    }
+
+    if (request_object.headers) {
+        options.headers = request_object.headers;
+    }
+
+    const response: Response = await fetch(request_object.path, options);
+    return await response.json();
+};
