@@ -1,128 +1,95 @@
-import { ContentRepository } from "../repository/ContentRepository";
-import { Callback } from "../../domain/callback/Callback";
+import { execute as contentRepository } from "../repository/ContentRepository";
+import { execute as callback } from "../../domain/callback/Callback";
 import { ResponseDTO } from "../dto/ResponseDTO";
-import { parser } from "../../application/variable/Parser";
 import { cache } from "../../application/variable/Cache";
 import { loaderInfoMap } from "../../application/variable/LoaderInfoMap";
 import type { LoaderInfo } from "@next2d/display";
-
-interface Object {
-    type: string;
-    name: string;
-    path: string;
-    cache?: boolean;
-    callback?: string | string[];
-    method?: string;
-    body?: object;
-    headers?: HeadersInit;
-}
+import type { RequestImpl } from "src/interface/RequestImpl";
 
 /**
- * NoCodeToolで制作したJSON取得時のロジッククラス
- * Logic class for JSON acquisition produced by NoCodeTool
+ * @description RepositoryからJSONを取得して、configのcallbackがあれば実行
+ *              キャッシュ設定がOnの時はJSONをキャッシュにセット
+ *              Get JSON from Repository and run config callback if any.
+ *              If cache setting is On, set JSON to cache.
  *
- * @class
- * @memberof infrastructure.service
+ * @param  {object} request_object
+ * @return {Promise<ResponseDTO>}
+ * @method
+ * @public
  */
-export class ContentService
+export const execute = async (request_object: RequestImpl): Promise<ResponseDTO> =>
 {
-    private _$repository: ContentRepository;
-    private _$callback: Callback;
-
-    /**
-     * @constructor
-     * @public
-     */
-    constructor ()
-    {
-        /**
-         * @type {ContentRepository}
-         * @private
-         */
-        this._$repository = new ContentRepository();
-
-        /**
-         * @type {Callback}
-         * @private
-         */
-        this._$callback = new Callback();
+    if (!request_object.name) {
+        throw new Error("`name` must be set for content requests.");
     }
 
     /**
-     * @description RepositoryからJSONを取得して、configのcallbackがあれば実行
-     *              キャッシュ設定がOnの時はJSONをキャッシュにセット
-     *              Get JSON from Repository and run config callback if any.
-     *              If cache setting is On, set JSON to cache.
-     *
-     * @param  {object} object
-     * @return {Promise<ResponseDTO>}
-     * @method
-     * @public
+     * キャッシュを利用する場合はキャッシュデータをチェック
+     * Check cache data if cache is used
      */
-    execute (object: Object): Promise<ResponseDTO>
-    {
-        /**
-         * キャッシュを利用する場合はキャッシュデータをチェック
-         * Check cache data if cache is used
-         */
-        if (object.cache && object.name) {
+    if (request_object.cache) {
 
-            const name: string = parser.execute(object.name);
-            if (cache.size && cache.has(name)) {
+        if (cache.size && cache.has(request_object.name)) {
 
-                const value: any = cache.get(name);
+            const value: any = cache.get(request_object.name);
 
+            /**
+             * コールバック設定があれば実行
+             * Execute callback settings if any.
+             */
+            if (request_object.callback) {
                 const promises: Promise<Awaited<any>[]|void>[] = [];
-                if (object.callback) {
-                    promises.push(this._$callback.execute(
-                        object.callback, value
-                    ));
-                }
+                promises.push(callback(
+                    request_object.callback, value
+                ));
 
-                return Promise
-                    .all(promises)
-                    .then((): ResponseDTO =>
-                    {
-                        return new ResponseDTO(name, value);
-                    });
+                await Promise.all(promises);
+            }
+
+            return new ResponseDTO(request_object.name, value);
+        }
+    }
+
+    /**
+     * 指定のコンテンツデータを取得
+     * Obtain specified content data
+     */
+    const content = await contentRepository(request_object);
+
+    /**
+     * キャッシュ設定がonならキャッシュに登録
+     * If the cache setting is on, register it in the cache.
+     */
+    if (request_object.cache) {
+        cache.set(request_object.name, content);
+    }
+
+    /**
+     * Animation Toolで設定したシンボルをマップに登録
+     * Register the symbols set by Animation Tool to the map
+     */
+    const loaderInfo: LoaderInfo = content._$loaderInfo as NonNullable<LoaderInfo>;
+    if (loaderInfo._$data) {
+        const symbols: Map<string, any> = loaderInfo._$data.symbols;
+        if (symbols.size) {
+            for (const name of symbols.keys()) {
+                loaderInfoMap.set(name, loaderInfo);
             }
         }
-
-        return this
-            ._$repository
-            .execute(object)
-            .then((content: any) =>
-            {
-                const name: string = parser.execute(object.name);
-                if (object.cache && object.name) {
-                    cache.set(name, content);
-                }
-
-                const loaderInfo: LoaderInfo = content._$loaderInfo;
-
-                // DisplayObjectContainer
-                if (loaderInfo._$data) {
-                    const symbols: Map<string, any> = loaderInfo._$data.symbols;
-                    if (symbols.size) {
-                        for (const name of symbols.keys()) {
-                            loaderInfoMap.set(name, loaderInfo);
-                        }
-                    }
-                }
-
-                const promises: Promise<Awaited<any>[]|void>[] = [];
-                if (object.callback) {
-                    promises.push(this._$callback.execute(
-                        object.callback, content
-                    ));
-                }
-
-                return Promise
-                    .all(promises)
-                    .then((): ResponseDTO =>
-                    {
-                        return new ResponseDTO(name, content);
-                    });
-            });
     }
-}
+
+    /**
+     * コールバック設定があれば実行
+     * Execute callback settings if any.
+     */
+    if (request_object.callback) {
+        const promises: Promise<Awaited<any>[]|void>[] = [];
+        promises.push(callback(
+            request_object.callback, content
+        ));
+
+        await Promise.all(promises);
+    }
+
+    return new ResponseDTO(request_object.name, content);
+};
