@@ -1,116 +1,63 @@
-import { CustomRepository } from "../repository/CustomRepository";
-import { Callback } from "../../domain/callback/Callback";
+import type { RequestImpl } from "src/interface/RequestImpl";
 import { ResponseDTO } from "../dto/ResponseDTO";
-import { parser } from "../../application/variable/Parser";
 import { cache } from "../../application/variable/Cache";
-
-interface Object {
-    type: string;
-    name: string;
-    path: string;
-    cache?: boolean;
-    class: string;
-    access: string;
-    method: string;
-    callback?: string | string[];
-    body?: object;
-    headers?: HeadersInit;
-}
+import { execute as callback } from "../../domain/callback/Callback";
+import { execute as customRepository } from "../repository/CustomRepository";
 
 /**
- * JSON取得時のロジッククラス
- * Logic class for JSON acquisition
+ * @description Repositoryから外部データを取得して、configのcallbackがあれば実行
+ *              キャッシュ設定がOnの時はJSONをキャッシュにセット
+ *              Retrieve external data from Repository and run config callback if any.
+ *              If cache setting is On, set JSON to cache.
  *
- * @class
- * @memberof infrastructure.service
+ * @param  {object} request_object
+ * @return {Promise<ResponseDTO>}
+ * @method
+ * @public
  */
-export class CustomService
+export const execute = async (request_object: RequestImpl): Promise<ResponseDTO> =>
 {
-    private readonly _$repository: CustomRepository;
-    private readonly _$callback: Callback;
-
-    /**
-     * @constructor
-     * @public
-     */
-    constructor ()
-    {
-        /**
-         * @type {CustomRepository}
-         * @private
-         */
-        this._$repository = new CustomRepository();
-
-        /**
-         * @type {Callback}
-         * @private
-         */
-        this._$callback = new Callback();
+    if (!request_object.name) {
+        throw new Error("`name` must be set for custom requests.");
     }
 
     /**
-     * @description Repositoryから外部データを取得して、configのcallbackがあれば実行
-     *              キャッシュ設定がOnの時はJSONをキャッシュにセット
-     *              Retrieve external data from Repository and run config callback if any.
-     *              If cache setting is On, set JSON to cache.
-     *
-     * @param  {object} object
-     * @return {Promise<ResponseDTO>}
-     * @method
-     * @public
+     * キャッシュを利用する場合はキャッシュデータをチェック
+     * Check cache data if cache is used
      */
-    execute (object: Object): Promise<ResponseDTO>
-    {
-        /**
-         * キャッシュを利用する場合はキャッシュデータをチェック
-         * Check cache data if cache is used
-         */
-        if (object.cache && object.name) {
+    if (request_object.cache) {
 
-            const name: string = parser.execute(object.name);
-            if (cache.size && cache.has(name)) {
+        if (cache.size && cache.has(request_object.name)) {
 
-                const value: any = cache.get(name);
+            const value: any = cache.get(request_object.name);
 
+            if (request_object.callback) {
                 const promises: Promise<Awaited<any>[]|void>[] = [];
-                if (object.callback) {
-                    promises.push(this._$callback.execute(
-                        object.callback, value
-                    ));
-                }
+                promises.push(callback(
+                    request_object.callback, value
+                ));
 
-                return Promise
-                    .all(promises)
-                    .then((): ResponseDTO =>
-                    {
-                        return new ResponseDTO(name, value);
-                    });
+                await Promise.all(promises);
             }
+
+            return new ResponseDTO(request_object.name, value);
         }
-
-        return this
-            ._$repository
-            .execute(object)
-            .then((response: any) =>
-            {
-                const name:string = parser.execute(object.name);
-                if (object.cache && object.name) {
-                    cache.set(name, response);
-                }
-
-                const promises: Promise<Awaited<any>[]|void>[] = [];
-                if (object.callback) {
-                    promises.push(this._$callback.execute(
-                        object.callback, response
-                    ));
-                }
-
-                return Promise
-                    .all(promises)
-                    .then((): ResponseDTO =>
-                    {
-                        return new ResponseDTO(name, response);
-                    });
-            });
     }
-}
+
+    const response: any = await customRepository(request_object);
+
+    if (request_object.cache) {
+        cache.set(request_object.name, response);
+    }
+
+    if (request_object.callback) {
+        const promises: Promise<Awaited<any>[]|void>[] = [];
+        promises.push(callback(
+            request_object.callback, response
+        ));
+
+        await Promise.all(promises);
+    }
+
+    return new ResponseDTO(request_object.name, response);
+};
