@@ -1,6 +1,5 @@
 import { StageImpl } from "./StageImpl";
 import { RoutingImpl } from "./RoutingImpl";
-import { LoadingImpl } from "./LoadingImpl";
 import { GotoViewImpl } from "./GotoViewImpl";
 
 interface BaseConfigImpl {
@@ -14,6 +13,8 @@ export interface ConfigImpl extends BaseConfigImpl {
         [key: string]: RoutingImpl
     };
     spa: boolean;
-    loading?: LoadingImpl;
+    loading?: {
+        callback: string;
+    };
     gotoView?: GotoViewImpl;
 }
